// FrontendApplication/src/app/enums/GameState.ts
export enum GameState {
  CREATED = 'CREATED',
  ONGOING = 'ONGOING',
  CHECK = 'CHECK',
  CHECKMATE = 'CHECKMATE',
  STALEMATE = 'STALEMATE',
  DRAW_BY_INSUFFICIENT_MATERIAL = 'DRAW_BY_INSUFFICIENT_MATERIAL',
  DRAW_BY_FIFTY_MOVES_RULE = 'DRAW_BY_FIFTY_MOVES_RULE',
  DRAW_BY_THREEFOLD_REPETITION = 'DRAW_BY_THREEFOLD_REPETITION',
  RESIGNATION = 'RESIGNATION',
  TIMEOUT = 'TIMEOUT',
  ABANDONED = 'ABANDONED',
  DRAW_AGREEMENT = 'DRAW_AGREEMENT',
  ERROR = 'ERROR'
}